requirements.txt
box2d-py==2.3.5
gym==0.26.2
h5py==3.7.0
imageio==2.22.4
ipykernel==6.16.0
ipython==8.5.0
jupyter==1.0.0
keras==2.11.0
matplotlib==3.6.1
notebook==6.5.1
numpy==1.23.4
opencv-python==4.6.0.66
Pillow==9.2.0
pygame==2.1.0
swig==4.1.0
tensorflow==2.11.0
torch==1.13.0
torchvision==0.14.0
